backend/requirements.txt
fastapi==0.87.0
starlette==0.21.0
pydantic==1.10.2
anyio==3.6.2
typing_extensions==4.4.0
sniffio==1.3.0
idna==3.4
sqlalchemy==1.4.44
greenlet==2.0.1
psycopg2_binary==2.9.5
passlib==1.7.4
bcrypt==4.0.1
uvicorn==0.20.0
fastapi-pagination==0.10.0
sqlalchemy-utils==0.38.3
alembic==1.8.1
MarkupSafe==2.1.1
python_jose==3.3.0
rsa==4.9
ecdsa==0.18.0
pyasn1==0.4.8
cryptography==38.0.3
cffi==1.15.1
six==1.16.0
pycparser==2.21
